🎨 프로필 생성 뷰 통합
